change price type in product

+++ migrations/1628375624656_create-table-users.js
@@ -21,7 +21,7 @@ exports.up = (pgm) => {
       notNull: true,
     },
     price: {
-      type: 'integer',
+      type: 'float8',
       notNull: true,
     },
     on_sell: {
